move v2 entrypoint to __main__
